requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiokafka==0.12.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
arrow==1.4.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.4.0
Authlib==1.6.5
boto3==1.40.58
boto3-stubs==1.40.58
botocore==1.40.58
botocore-stubs==1.40.59
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
circuitbreaker==2.1.3
click==8.0.4
cryptography==44.0.3
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
frozenlist==1.8.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.11.1
jmespath==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langchain-weaviate==0.0.5
langgraph==1.0.0
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.0
langgraph-sdk==0.2.9
langsmith==0.4.37
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.4
oci==2.161.1
oci-cli==3.68.1
openai==2.5.0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pip-chill==1.0.3
prompt-toolkit==3.0.43
propcache==0.4.1
protobuf==6.33.0
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyOpenSSL==24.3.0
pypdf==6.1.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
s3transfer==0.14.0
setuptools==80.9.0
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
tenacity==9.1.2
terminaltables==3.1.10
tiktoken==0.12.0
tqdm==4.67.1
types-awscrt==0.28.2
types-s3transfer==0.14.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
validators==0.35.0
wcwidth==0.2.14
weaviate-client==4.17.0
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
